Changed function invokeWebhook to allow for additional settings (#487)
* Changed function invokeWebhook to allow for additional settings

The invokeWebhook function was changed so that the following settings can be used in the settings.json file:
webhook_type - set the name of the entity that contains the type (defaults to type)
webhook_data - set the name of the entity that contains the data (defaults to data)
webhook_header_name - add a header (for instance for authorization) (defaults to none)
webhook_header_contents - set the contents of the additional header.

The reason for this change is that I want to point this webhook to Splunk. Splunk requires an event entity in the payload in stead of a data entity and the type entity is unknown as well. In addition, authentication is done through the header "Authentication".

I have programmed it such that it is backwards compatible with the previous version.

* Changed new property and variable names to camelCase