Restrict/allow concurrent '...all' actions
Concurrent clipall/sayall can leave the system in an undefined grouped state.
This feature implements a check within all-player-announcements. Calls to the helper will start a timer and, if restricted, subsequent calls will be ignored while the timer is active.
This allows time for the original state to be restored before the next clipall/sayall action.

Global variables are possibly not the best implementation, but none of the existing helper files require settings.js so this change did not introduce a require to all-player-announcements.