Enhanced debugging features: 1. Log request uri in server request handler for log level debug. I prefer to log this level info in mny fork because it enables quick diagnosis when my Alexa speech skill has misheard what I said. But doing it on debug level is more consistent with the original program.
2. Add /loglevel get request which implements ability to
change the process NODE_LOG_LEVEL (one of trace, debug,
info, warn, or error (and off which is a synonym for error)).
Document the command in static doc content.

This is very useful when the code is running in a container. Especially
if you add it to your alexa skill as I have.